order implementation

+++ domain/src/main/java/org/dbs/shop/domain/Item.java
@@ -1,7 +1,13 @@
 package org.dbs.shop.domain;
 
+import java.math.BigDecimal;
+
 public class Item {
+
 	private String productName;
-	private float productPrice;
+
+	private BigDecimal productPrice;
+
 	private int quantity;
+
 }
